+++ HaselDebug/Tabs/AgentMapEventMarkersTab.cs
@@ -4,6 +4,7 @@ using FFXIVClientStructs.FFXIV.Client.UI.Agent;
 using HaselDebug.Abstracts;
 using HaselDebug.Interfaces;
 using ImGuiNET;
+using Lumina.Text.ReadOnly;
 
 namespace HaselDebug.Tabs;
 
@@ -18,25 +19,27 @@ public unsafe partial class AgentMapEventMarkersTab : DebugTab
     {
         var agent = AgentMap.Instance();
 
-        using var table = ImRaii.Table("AgentMapEventMarkersTable", 6, ImGuiTableFlags.RowBg | ImGuiTableFlags.Borders | ImGuiTableFlags.ScrollY | ImGuiTableFlags.NoSavedSettings);
+        using var table = ImRaii.Table("AgentMapEventMarkersTable", 7, ImGuiTableFlags.RowBg | ImGuiTableFlags.Borders | ImGuiTableFlags.ScrollY | ImGuiTableFlags.NoSavedSettings);
         if (!table) return;
 
         ImGui.TableSetupColumn("Icon", ImGuiTableColumnFlags.WidthFixed, 100);
         ImGui.TableSetupColumn("LevelId", ImGuiTableColumnFlags.WidthFixed, 100);
         ImGui.TableSetupColumn("ObjectiveId", ImGuiTableColumnFlags.WidthFixed, 100);
         ImGui.TableSetupColumn("MapId", ImGuiTableColumnFlags.WidthFixed, 100);
+        ImGui.TableSetupColumn("Radius", ImGuiTableColumnFlags.WidthFixed, 100);
         ImGui.TableSetupColumn("TerritoryTypeId", ImGuiTableColumnFlags.WidthFixed, 100);
         ImGui.TableSetupColumn("TooltipString");
         ImGui.TableSetupScrollFreeze(6, 1);
         ImGui.TableHeadersRow();
 
-        foreach (var marker in agent->EventMarkers)
+        foreach (ref var marker in agent->EventMarkers)
         {
             ImGui.TableNextRow();
 
             ImGui.TableNextColumn(); // Icon
-            if (_textureProvider.GetFromGameIcon(marker.IconId).TryGetWrap(out var tex, out var _))
+            if (_textureProvider.TryGetFromGameIcon(marker.IconId, out var sharedTex) && sharedTex.TryGetWrap(out var tex, out var _))
                 ImGui.Image(tex.ImGuiHandle, new(ImGui.GetTextLineHeight()));
+
             ImGui.SameLine();
             ImGui.TextUnformatted(marker.IconId.ToString());
 
@@ -49,11 +52,15 @@ public unsafe partial class AgentMapEventMarkersTab : DebugTab
             ImGui.TableNextColumn(); // MapId
             ImGui.TextUnformatted(marker.MapId.ToString());
 
+            ImGui.TableNextColumn(); // Radius
+            ImGui.TextUnformatted(marker.Radius.ToString());
+
             ImGui.TableNextColumn(); // TerritoryTypeId
             ImGui.TextUnformatted(marker.TerritoryTypeId.ToString());
 
             ImGui.TableNextColumn(); // TooltipString
-            ImGui.TextUnformatted(marker.TooltipString->ToString());
+            if (marker.TooltipString != null && marker.TooltipString->StringPtr.Value != null)
+                ImGui.TextUnformatted(new ReadOnlySeStringSpan(marker.TooltipString->StringPtr.Value).ExtractText());
         }
     }
 }