+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+- **Fixed:** Agent Map Event Markers table broke when the icon wasn't found.
 - **Fixed:** Added the missing GameObject -> Companion type redirect.
 - **Fixed:** Object Table tab now sorts by index.
 - **Updated:** ClientStructs now at [c24c833a](https://github.com/aers/FFXIVClientStructs/tree/c24c833a) ([compare](https://github.com/aers/FFXIVClientStructs/compare/a625ce4d..c24c833a)).